requirements.txt
google-generativeai==0.3.0
websockets==12.0
asyncio-mqtt==0.13.0
python-dotenv==1.0.0